Rename LoggerFactory into LoggerManager
